Move Population struct and equi pop set-up to new files
